refactor: :truck: Rename system/dev directory to system/vfs
