+++ src/system/vfs/usd.cpp
@@ -1,4 +1,4 @@
-#include "system/dev/usd.hpp"
+#include "system/vfs/usd.hpp"
 
 #include "v5_api.h"
 #include "v5_apitypes.h"
